Crate spin_variables

The runtime configuration for the variables factor used in the Spin CLI.

Structs

AzureKeyVaultProvider

A provider that fetches variables from Azure Key Vault.

AzureKeyVaultVariablesConfig

Azure KeyVault runtime config literal options for authentication

EnvVariablesConfig

Configuration for the environment variables provider.

EnvVariablesProvider

A [Provider] that uses environment variables.

StaticVariablesProvider

A [Provider] that reads variables from an static map.

VaultVariablesProvider

A config Provider that uses HashiCorp Vault.

Enums

AzureAuthorityHost

AzureKeyVaultAuthOptions

Azure Cosmos Key / Value enumeration for the possible authentication options

VariableProviderConfiguration

A runtime configuration used in the Spin CLI for one type of variable provider.

Functions

runtime_config_from_toml

Resolves a runtime configuration for the variables factor from a TOML table.